타입 분리
